requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-SocketIO==2.9.2
geojson==2.1.0
google-auth==1.1.0
greenlet==0.4.12
gspread==0.6.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyasn1==0.3.4
pyasn1-modules==0.1.4
python-engineio==1.7.0
python-socketio==1.8.0
requests==2.18.4
rsa==3.4.2
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
wincertstore==0.2
